requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.33.0
django-rest-auth==0.9.2
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
idna==2.6
oauthlib==2.0.5
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
